feat(local_agent): add golden tests for chat page and llm settings
- Created chat_page_golden_test.dart covering Initial and Conversation states.
- Created llm_settings_page_golden_test.dart covering Gemini and Local LLM configurations.
- Created local_agent_widgets_golden_test.dart for MessageBubble and VoiceInputButton.
- Removed redundant and misplaced local_agent_page_golden_test.dart.
- Updated golden reference images.
- Ensured all golden tests pass and follow project standards using wrapWithMaterial and setupGoldenTest.
